A concrete usage of the VolumeConstrainedContainer
Example of an Item (using the Common/Item class) that is
a container of objects, depending on your item system, this object
holds its inventory independently and could be dropped on the floor
or otherwise removed from a players body.

That is, the Inventory is bound into the object itself, not the player
or creature holding it.