Test for the VolumeConstrainedContainer
Just assert that it won't add beyond it's total capacity.

Weight limits are difficult on a bag, because it's more likely to be
the character or creature carrying the bag will hit it's limit of
weight before the bag itself, but we posit 60KG of weight as a limit
the same to illustrate the effect.

A system using this is practice would have to devise its own method for
deriving the total weight capacity of a given container.